Add changes for building
