--- src/main/java/com/ossez/lang/tutorial/models/ListNode.java
@@ -1,16 +1,0 @@
-package com.ossez.lang.tutorial.models;
-
-/**
- * 
- * @author YuCheng
- *
- */
-public class ListNode {
-	public  int val;
-	public  ListNode next;
-
-	public ListNode(int x) {
-		val = x;
-		next = null;
-	}
-}
