--- src/main/java/com/ossez/lang/tutorial/models/TreeNode.java
@@ -1,16 +1,0 @@
-package com.ossez.lang.tutorial.models;
-
-/**
- * 
- * @author YuCheng
- *
- */
-public class TreeNode {
-	public int val;
-	public TreeNode left, right;
-
-	public TreeNode(int val) {
-		this.val = val;
-		this.left = this.right = null;
-	}
-}
