+++ toolkits/codebank/pom.xml
@@ -41,15 +41,15 @@
     </dependencies>
 
     <build>
-        <finalName>discourse</finalName>
+        <finalName>codebank</finalName>
         <plugins>
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-compiler-plugin</artifactId>
                 <version>${maven-compiler-plugin.version}</version>
                 <configuration>
-                    <source>${java.version}</source>
-                    <target>${java.version}</target>
+                    <source>11</source>
+                    <target>11</target>
                 </configuration>
             </plugin>
         </plugins>
