+++ toolkits/codebank/src/main/java/com/ossez/toolkits/codebank/common/model/ListNode.java
@@ -1,0 +1,16 @@
+package com.ossez.toolkits.codebank.common.model;
+
+/**
+ * 
+ * @author YuCheng
+ *
+ */
+public class ListNode {
+	public  int val;
+	public  ListNode next;
+
+	public ListNode(int x) {
+		val = x;
+		next = null;
+	}
+}
