+++ toolkits/codebank/src/test/java/com/ossez/toolkits/codebank/tests/lintcode/LintCode1480DotProductTest.java
